Models by create date
Partial progress toward #20

Does not resolve it because (amongst other reasons) we ought to allow
searching for term use over time (e.g. can we compare the growth of models
for one term to those for another? It would be nice to be able to look at both
the totals and the short-term numbers.)