Enhanced search and bugfixes on some views
Search now finds authors, simulation environments, currents, concepts,
regions, receptors, transmitters, genes in addition to what it found before.
Progress toward #5 and resolves #34

As part of the process, I discovered that regions and genes were not being
found correctly on the modellist page; this is fixed.

This fix might resolve the bug @tommorse mentioned in #52